requirements.txt
keyboard
pynput
requests
cdifflib
pyyaml
psutil
pyWin32